requirements.txt
streamlit==1.45.1
yfinance==0.2.61
pandas==2.2.3
numpy==2.2.5
plotly==5.19.0
nltk==3.8.1
textblob==0.17.1
tweepy==4.14.0
praw==7.7.1
newsapi-python==0.2.7
python-dotenv==1.0.1
sqlalchemy==2.0.28
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2 